Move implementation into header
